+++ methods/get-version.js
@@ -3,10 +3,6 @@ const versionInfo = require('../public/version.json');
 
 Meteor.methods({
   getVersion: function() {
-    // if (this.isSimulation) return;
-    console.log(versionInfo);
     return versionInfo.version;
-  // if (Meteor.isServer) {
-  //   }
   },
 });